+++ src/main/java/uk/ac/ebi/jmzml/model/mzml/SpectrumList.java
@@ -5,7 +5,6 @@ import uk.ac.ebi.jmzml.xml.jaxb.adapters.IdRefAdapter;
 import javax.xml.bind.annotation.*;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -43,7 +42,7 @@ public class SpectrumList
     protected List<Spectrum> spectrum;
     @XmlAttribute(required = true)
     @XmlSchemaType(name = "nonNegativeInteger")
-    protected BigInteger count;
+    protected Integer count;
     @XmlAttribute(required = true)
     @XmlJavaTypeAdapter(IdRefAdapter.class)
     @XmlSchemaType(name = "IDREF")
@@ -82,9 +81,9 @@ public class SpectrumList
      * Gets the value of the count property.
      *
      * @return possible object is
-     *         {@link BigInteger }
+     *         {@link Integer }
      */
-    public BigInteger getCount() {
+    public Integer getCount() {
         return count;
     }
 
@@ -92,9 +91,9 @@ public class SpectrumList
      * Sets the value of the count property.
      *
      * @param value allowed object is
-     *              {@link BigInteger }
+     *              {@link Integer }
      */
-    public void setCount(BigInteger value) {
+    public void setCount(Integer value) {
         this.count = value;
     }
